Codec should be independent from circe, so one can use YAML or other codec
